+++ jobs/clean_ascwds_worker_data.py
@@ -8,7 +8,7 @@ from utils.column_names.raw_data_files.ascwds_worker_columns import PartitionKey
 from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
     AscwdsWorkerCleanedColumns as AWKClean,
 )
-from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned_values import (
+from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
 )
 from utils.value_labels.ascwds_worker.worker_label_dictionary import (
