+++ jobs/clean_ascwds_workplace_data.py
@@ -9,7 +9,7 @@ import utils.cleaning_utils as cUtils
 from utils.column_names.raw_data_files.ascwds_workplace_columns import (
     PartitionKeys as Keys,
 )
-from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned_values import (
+from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
 )
 from utils.scale_variable_limits import AscwdsScaleVariableLimits
