+++ tests/test_file_schemas.py
@@ -59,7 +59,7 @@ from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
 from utils.column_names.cleaned_data_files.cqc_location_cleaned_values import (
     CqcLocationCleanedColumns as CQCLClean,
 )
-from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned_values import (
+from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
 )
 from utils.column_names.cleaned_data_files.ascwds_worker_cleaned import (
